requirements.txt
python-telegram-bot>=20.0
groq>=0.4.0
requests>=2.28.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
flask